+++ src/dtos/TimeSnippet.ts
